refactor(travelagency): Services sous Infrastructure + suppression du doublon obsolète (Closes #7150)
