+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## [Unreleased]
+- **refactor(travelagency): placement couche Application + suppression d'un doublon obsolète (Closes #7150).** `TravelReportService` (utilisé par `TravelReportController`) déplacé sous `Infrastructure/Services/` ; `TravelAdvertPricingService` : la copie `Infrastructure` (`computePrice`, colonnes `type_id`/`position_id` **inexistantes** au schéma canonique #6109 `2026_08_30_001545`) a été **supprimée** et la version conforme au schéma (`quote()`, `advert_type_id`/`advert_position_id`/`currency`) déplacée sous `Infrastructure/Services/` avec note de canonisation. Aucun consommateur des deux services (logique inline dans `SubmitTravelAdvertAction`) : suppression sans impact runtime. Gardes : `check-application-layer-placement.sh` (0 violation TravelAgency), `check-layer-purity.sh` ✅.
 - **docs(ops): brief de décision desktop BC-08 Accounting (P06 §2, #7055) + runbook de rotation des jetons (#7065).** `docs/architecture/DESKTOP_BC_ACCOUNTING_DECISION.md` : réponses motivées aux 5 points P06 §2, recommandation GO conditionnel phase M1 (builds non signés + recette interne) et NO-GO téléchargement public avant GA ; décision finale PM. `docs/ops/JETONS_ROTATION_RUNBOOK.md` : procédure de rotation GitHub/Vercel/Render/Cloudflare (inventaire, rotation avec vérification, contrôle d'usage avant révocation, checklist, prévention) — aucun secret dans le document, exécution à la main du PM.
 - **docs(tooling): index docs complet (54/54), catalogue d'événements dédupliqué + garde d'unicité, garde design inter-surfaces (Closes #7147, #7148, #7149).** `docs/README.md` référence désormais les 54 dossiers de `docs/` (+ section « Domaines métier & guides ») ; `dev-hub/governance/event-catalogue.json` déduplonné (25→23 entrées, `AttendanceRecorded`/`FuelStationAlert`) et `check-event-catalogue.py` échoue sur tout doublon ; `EVENT_CATALOGUE.md` complété (événements BC-28 `catalog.inquiry_received`/`erased`) ; `check-design-token-sync.py` compare désormais les palettes **partagées web ↔ admin** (la dérive cyan de #7129 était invisible pour la garde). Tests négatifs exécutés. Aucun changement de comportement runtime.
 - **docs(ops): DR realigné sur le fail-loud effectif des backups — suppression du `::notice::` fantôme (#6836).** `docs/ops/DR.md` documentait encore l'ancien comportement (`::notice::` + succès) alors que `.github/workflows/database-backup.yml` échoue volontairement depuis #6892 : §3.1 et §7 décrivent désormais l'échec `::error::` + `exit 1` listant les secrets manquants (garde #6836, convention anti-faux-vert #6831) ; statut d'en-tête requalifié 🟡 et note d'état datée (aucun dump tant que `BACKUP_S3_BUCKET`/`AWS_*` manquent → RPO §2 non tenu), pour ne plus laisser croire à une sauvegarde en place. Aucun changement de workflow ni de comportement runtime.
